Backend/dev_media/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mysqlclient
pillow